backend/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
azure-common==1.1.27
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
black
certifi==2020.12.5
cffi==1.14.4
cfgv==3.2.0
chardet==3.0.4
colorama==0.4.4
coreapi==2.3.3
# cryptography==3.2.1
coverage==5.5
distlib==0.3.1
celery-progress==0.1.1
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
djangorestframework-simplejwt
django-storages==1.11.1
drf-yasg==1.20.0
filelock==3.0.12
gunicorn==20.0.4
identify==1.5.10
#idna==2.10
importlib-metadata==3.1.1
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
msal==1.7.0
nodeenv==1.5.0
pre-commit==2.9.3
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
pylint==2.6.0
pytz==2020.1
PyYAML==5.3.1
# requests==2.0.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
# urllib3==1.25.11
virtualenv==20.2.2
wrapt==1.12.1
zipp==3.4.0
django-filter==2.4.0
celery==5.1.2
redis==3.5.3
pandas-profiling
#pandas==0.25.3
# Cython